# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daton Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/departments:
    get:
      operationId: listDepartments
      tags: [organizations]
      summary: List departments
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [organizations]
      summary: Create department
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentBody"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /organizations/{orgId}/departments/{deptId}:
    patch:
      operationId: updateDepartment
      tags: [organizations]
      summary: Update department
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: deptId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentBody"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [organizations]
      summary: Delete department
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: deptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Department deleted
  /organizations/{orgId}/positions:
    get:
      operationId: listPositions
      tags: [organizations]
      summary: List positions
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
    post:
      operationId: createPosition
      tags: [organizations]
      summary: Create position
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePositionBody"
      responses:
        "201":
          description: Position created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /organizations/{orgId}/positions/{posId}:
    patch:
      operationId: updatePosition
      tags: [organizations]
      summary: Update position
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: posId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePositionBody"
      responses:
        "200":
          description: Updated position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
    delete:
      operationId: deletePosition
      tags: [organizations]
      summary: Delete position
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: posId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Position deleted
...
components:
  schemas:
...
    Department:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
    CreateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
    UpdateDepartmentBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    Position:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        education:
          type: string
          nullable: true
        experience:
          type: string
          nullable: true
        requirements:
          type: string
          nullable: true
        responsibilities:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
    CreatePositionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        education:
          type: string
        experience:
          type: string
        requirements:
          type: string
        responsibilities:
          type: string
      required:
        - name
    UpdatePositionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        education:
          type: string
        experience:
          type: string
        requirements:
          type: string
        responsibilities:
          type: string
...